@@ -15,7 +15,7 @@ class FadingText extends StatefulWidget {
   /// Text to animate
   final String text;
   /// Custom text style. If not specified, uses the default style.
-  final TextStyle style;
+  final TextStyle? style;
 
   /// Creates a fading continuous animation.
   ///
@@ -29,7 +29,7 @@ class FadingText extends StatefulWidget {
 
 class _FadingTextState extends State<FadingText> with TickerProviderStateMixin {
   final _characters = <MapEntry<String, Animation>>[];
-  AnimationController _controller;
+  late AnimationController _controller;
 
   @override
   void initState() {
@@ -74,7 +74,7 @@ class _FadingTextState extends State<FadingText> with TickerProviderStateMixin {
       children: _characters
           .map(
             (entry) => FadeTransition(
-                  opacity: entry.value,
+                  opacity: entry.value as Animation<double>,
                   child: Text(entry.key, style: widget.style),
                 ),
           )
@@ -101,7 +101,7 @@ class JumpingText extends StatelessWidget {
   final String text;
   final Offset begin = Offset(0.0, 0.0);
   final Offset end;
-  final TextStyle style;
+  final TextStyle? style;
 
   /// Creates a jumping text widget.
   ///
@@ -136,7 +136,7 @@ class ScalingText extends StatelessWidget {
   final String text;
   final double begin = 1.0;
   final double end;
-  final TextStyle style;
+  final TextStyle? style;
 
   /// Creates a jumping text widget.
   ///
